Add a new boolean option, relativeToWrapper.
This addition originates from the need to let movement be relative to a wrapper, rather than scroll position of a wrapper or the scroll position of the window.
We'll use the wrappers vertical offset, and subtract it from the scrollposition, to enable an effect, that feels like the element is relative to its wrapper instead.

Simply a boolean, telling us whether or not it should be relative to the container selected in the wrapper option.